changed to floats, added paralleograms

+++ include/CheckeredTexture.h
@@ -8,9 +8,9 @@ namespace rt{
 
 class CheckeredTexture: public ITexture{
 public:
-    CheckeredTexture(double scale, TextureHandle even, TextureHandle odd): inverseScale{1.0/scale}, even{even}, odd{odd}{}
+    CheckeredTexture(float scale, TextureHandle even, TextureHandle odd): inverseScale{1.0f/scale}, even{even}, odd{odd}{}
 
-    glm::dvec3 color(double u, double v, glm::dvec3 point, const Raytracer& env) const override{
+    glm::dvec3 color(float u, float v, glm::vec3 point, const Raytracer& env) const override{
         int xInt = static_cast<int>(glm::floor(inverseScale * point.x));
         int yInt = static_cast<int>(glm::floor(inverseScale * point.y));
         int zInt = static_cast<int>(glm::floor(inverseScale * point.z));
@@ -21,7 +21,7 @@ public:
     }
 private:
 
-    double inverseScale;
+    float inverseScale;
     TextureHandle even;
     TextureHandle odd;
 };
