+++ include/ITexture.h
@@ -8,6 +8,6 @@ class ITexture {
 public:
     virtual ~ITexture() = default;
 
-    virtual glm::dvec3 color(double u, double v, glm::dvec3 point, const Raytracer& env) const = 0;
+    virtual glm::dvec3 color(float u, float v, glm::vec3 point, const Raytracer& env) const = 0;
 };
 }
