+++ include/MetalMaterial.h
@@ -10,11 +10,11 @@ class Raytracer;
 
 class MetalMaterial : public IMaterial{
 public:
-    MetalMaterial(TextureHandle albedo, double fuzz);
+    MetalMaterial(TextureHandle albedo, float fuzz);
     bool scatter(const Ray& ray, const HitRecord& rec, glm::dvec3& attenuation, Ray& scattered, const Raytracer& env) const override;
 
 private:
     TextureHandle albedo;
-    double fuzz;
+    float fuzz;
 };
 }
